♻️ refactor: monster-name 컴포넌트 위치 이동
- 비슷한 관심사 폴더 내 이동
