+++ src/app/(route)/(auth)/signup/page.tsx
@@ -11,9 +11,9 @@ import useFunnel from '@hooks/useFunnel';
 import MakeNickName from './components/MakeNickName';
 import { funnel, title } from './constant/funnel';
 import { UserInfoProvider } from './context/userInfo.context';
-import MonsterName from '../../(monster)/component/MosterName';
 import CustomizeMonster from '../../(monster)/monster/produce/components/CustomizeMonster';
 import MonsterLayout from '../../(monster)/monster/produce/components/layout/MonsterLayout';
+import MonsterName from '../../(monster)/monster/produce/components/MosterName';
 import SelectEmotion from '../../(monster)/monster/produce/components/SelectEmotion';
 import SelectStress from '../../(monster)/monster/produce/components/SelectStress';
 import WriteStory from '../../(monster)/monster/produce/components/WriteStory';
